Handle UB in bit_scan_forward

+++ include/eixx/util/common.hpp
@@ -74,7 +74,8 @@ int __inline__ log2(unsigned long n, uint8_t base = 2) {
     return n == 1 ? 0 : 1+log2(n/base, base); 
 }
 
-static __inline__ unsigned long bit_scan_forward(unsigned long v)
+/// Note, that bit_scan_forward(0) leads to UB
+static __inline__ int bit_scan_forward(unsigned long v)
 {
     return __builtin_ctzl(v);
 }
